scale and square like time evolution
Time evolution with Taylor method is sensitive to roundoff error.
To minimize this type of error we follow a scale and square like
method.  This method exploits the fact that the matrix exponential
of a matrix A (exp(A)) can be written as exp(A/m)^m for an integer
m. We select m such that the sum of absolute values of the
coefficient tensor definining the operator is 0.5 to ensure
convergence.  Likely this parameter can be optimized.

I've also reorganized the set of functions so there is less
code repetition.